Move competition index to root
