+++ apps/wodsmith-start/src/components/competition-sidebar.tsx
@@ -208,7 +208,7 @@ function CompetitionSidebarHeader() {
   return (
     <SidebarHeader className="h-14 flex-row items-center border-b px-3 group-data-[collapsible=icon]:px-2 group-data-[collapsible=icon]:justify-center">
       <Link
-        to="/compete"
+        to="/"
         className="flex items-center gap-2 min-w-0 group-data-[collapsible=icon]:hidden"
       >
         <img
@@ -226,7 +226,7 @@ function CompetitionSidebarHeader() {
         </h1>
       </Link>
       <Link
-        to="/compete"
+        to="/"
         className="hidden group-data-[collapsible=icon]:block"
       >
         <img
@@ -316,7 +316,7 @@ export function CompetitionSidebar({
           <SidebarTrigger className="-ml-1">
             <Menu className="h-5 w-5" />
           </SidebarTrigger>
-          <Link to="/compete" className="flex items-center gap-2">
+          <Link to="/" className="flex items-center gap-2">
             <img
               src="/wodsmith-logo-no-text.png"
               alt="wodsmith compete"
