+++ apps/wodsmith-start/src/components/landing/two-audiences.tsx
@@ -137,7 +137,7 @@ export function TwoAudiences({ session }: TwoAudiencesProps) {
 
             <div className="mt-auto pt-6 border-t border-border">
               <Button variant="outline" className="w-full" asChild>
-                <Link to="/compete">
+                <Link to="/">
                   Browse Competitions
                   <ArrowRight className="ml-2 h-4 w-4" />
                 </Link>
