+++ apps/wodsmith-start/src/routes/_protected/admin/teams/scaling/index.tsx
@@ -43,7 +43,7 @@ export const Route = createFileRoute("/_protected/admin/teams/scaling/")({
   component: AdminScalingPage,
   beforeLoad: async ({ context }) => {
     if (!context.hasWorkoutTracking) {
-      throw redirect({ to: "/compete" })
+      throw redirect({ to: "/" })
     }
   },
   loader: async ({ context }) => {
