+++ apps/wodsmith-start/src/routes/_protected/movements/index.tsx
@@ -16,7 +16,7 @@ export const Route = createFileRoute("/_protected/movements/")({
   component: MovementsPage,
   beforeLoad: async ({ context }) => {
     if (!context.hasWorkoutTracking) {
-      throw redirect({ to: "/compete" })
+      throw redirect({ to: "/" })
     }
   },
   validateSearch: (search: Record<string, unknown>) => ({
