+++ apps/wodsmith-start/src/routes/_protected/workouts/index.tsx
@@ -73,7 +73,7 @@ export const Route = createFileRoute("/_protected/workouts/")({
   component: WorkoutsPage,
   beforeLoad: async ({ context }) => {
     if (!context.hasWorkoutTracking) {
-      throw redirect({ to: "/compete" })
+      throw redirect({ to: "/" })
     }
   },
   validateSearch: (search: Record<string, unknown>): WorkoutsSearch => {
